requirements.txt
Package                  Version
------------------------ --------------------
absl-py                  2.3.1
apex                     0.1
Brotli                   1.1.0
cachetools               5.5.2
certifi                  2024.8.30
cffi                     1.17.0
charset-normalizer       3.4.0
colorama                 0.4.6
contourpy                1.1.1
cycler                   0.12.1
easydict                 1.13
einops                   0.8.0
faiss-gpu                1.7.2
filelock                 3.16.1
fonttools                4.53.1
fsspec                   2025.3.0
fvcore                   0.1.5.post20221221
google-auth              2.40.3
google-auth-oauthlib     1.0.0
grpcio                   1.70.0
h2                       4.1.0
h5py                     3.11.0
hf-xet                   1.1.9
hpack                    4.0.0
huggingface-hub          0.34.4
hyperframe               6.0.1
idna                     3.10
importlib_metadata       8.5.0
importlib_resources      6.4.5
iopath                   0.1.10
Jinja2                   3.1.6
joblib                   1.4.2
kiwisolver               1.4.5
llvmlite                 0.41.1
Markdown                 3.7
MarkupSafe               2.1.5
matplotlib               3.7.3
mpmath                   1.3.0
munkres                  1.1.4
networkx                 3.1
numba                    0.58.1
numpy                    1.24.4
nvidia-cublas-cu12       12.1.3.1
nvidia-cuda-cupti-cu12   12.1.105
nvidia-cuda-nvrtc-cu12   12.1.105
nvidia-cuda-runtime-cu12 12.1.105
nvidia-cudnn-cu12        9.1.0.70
nvidia-cufft-cu12        11.0.2.54
nvidia-curand-cu12       10.3.2.106
nvidia-cusolver-cu12     11.4.5.107
nvidia-cusparse-cu12     12.1.0.106
nvidia-nccl-cu12         2.20.5
nvidia-nvjitlink-cu12    12.9.86
nvidia-nvtx-cu12         12.1.105
oauthlib                 3.3.1
olefile                  0.47
opencv-python-headless   4.11.0.86
packaging                25.0
pandas                   2.0.3
patsy                    0.5.6
pickle5                  0.0.11
Pillow                   8.4.0
pip                      25.0.1
platformdirs             4.3.6
ply                      3.11
pooch                    1.8.2
portalocker              3.0.0
protobuf                 5.29.5
pyasn1                   0.6.1
pyasn1_modules           0.4.2
pycparser                2.22
pynndescent              0.5.13
pyparsing                3.1.4
PySocks                  1.7.1
python-dateutil          2.9.0
pytz                     2024.2
PyWavelets               1.3.0
PyYAML                   6.0.2
requests                 2.32.3
requests-oauthlib        2.0.0
rsa                      4.9.1
safetensors              0.5.3
scikit-learn             1.2.0
scipy                    1.10.1
seaborn                  0.13.2
setuptools               75.1.0
sip                      6.7.12
six                      1.16.0
statsmodels              0.14.1
sympy                    1.13.3
tabulate                 0.9.0
tensorboard              2.14.0
tensorboard-data-server  0.7.2
termcolor                2.4.0
thop                     0.1.1.post2209072238
threadpoolctl            3.5.0
timm                     1.0.19
toml                     0.10.2
tomli                    2.0.2
torch                    2.4.1
torchaudio               0.12.1
torchcam                 0.4.0
torchinfo                1.8.0
torchvision              0.13.1
tornado                  6.4.1
tqdm                     4.64.1
triton                   3.0.0
typing_extensions        4.12.2
tzdata                   2024.2
umap-learn               0.5.6
unicodedata2             15.1.0
urllib3                  2.2.3
Werkzeug                 3.0.6
wheel                    0.44.0
yacs                     0.1.8
zipp                     3.21.0
zstandard                0.23.0